Rough-in an endpoint that returns JSON.
