+++ auslib/web/public/client.py
@@ -1,13 +1,13 @@
 import logging
 import re
 import sys
-from functools import wraps
 
 from connexion import request
 from flask import current_app as app
 from flask import make_response
 
 from auslib.AUS import AUS, FAIL, SUCCEED
+from auslib.web.public.base import with_transaction
 from auslib.global_state import dbo
 
 try:
@@ -123,15 +123,6 @@ def extract_query_version(request_url):
     if match:
         version = int(match.group(1))
     return version
-
-
-def with_transaction(f):
-    @wraps(f)
-    def wrapper(*args, **kwargs):
-        with dbo.begin() as transaction:
-            return f(*args, transaction=transaction, **kwargs)
-
-    return wrapper
 
 
 @with_transaction
